requirements.txt
Django>=4.2
djangorestframework
requests
textblob
python-dotenv
django-environ
pytest
pytest-django
gunicorn
stripe
